# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /driver/destination-mode:
    get:
      tags: [Driver]
      operationId: getDriverDestinationMode
      summary: Get the driver's current destination mode state and quota
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Destination mode state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationModeState'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /driver/destination-mode/activate:
    post:
      tags: [Driver]
      operationId: activateDriverDestinationMode
      summary: Turn on destination mode toward an address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateDestinationModeRequest'
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationModeState'
        '400':
          description: Invalid input or feature disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Daily activation cap reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /driver/destination-mode/deactivate:
    post:
      tags: [Driver]
      operationId: deactivateDriverDestinationMode
      summary: Turn off destination mode
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationModeState'
  /driver/saved-places:
    get:
      tags: [Driver]
      operationId: listDriverSavedPlaces
      summary: List the driver's saved places (Home/Work) and recents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saved places + recents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverSavedPlacesResponse'
    post:
      tags: [Driver]
      operationId: upsertDriverSavedPlace
      summary: Save or upsert a Home/Work/recent place for the driver
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertDriverSavedPlaceRequest'
      responses:
        '200':
          description: Upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverSavedPlace'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /driver/saved-places/{id}:
    delete:
      tags: [Driver]
      operationId: deleteDriverSavedPlace
      summary: Delete a saved/recent place
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    DestinationModeConfig:
      type: object
      required: [enabled, maxPerDay, matchRadiusKm, corridorKm, autoDisableOnTrip, autoDisableMinutes]
      properties:
        enabled: { type: boolean, description: "Master feature flag from admin settings." }
        maxPerDay: { type: integer, minimum: 0, description: "How many activations the driver may make per 24h." }
        matchRadiusKm: { type: number, description: "Dropoff match radius around the destination." }
        corridorKm: { type: number, description: "Cross-track corridor width along the pickup→dropoff bearing." }
        autoDisableOnTrip: { type: boolean }
        autoDisableMinutes: { type: integer, minimum: 0 }
    DestinationModeActive:
      type: object
      required: [id, address, lat, lng, label, activatedAt]
      properties:
        id: { type: string, format: uuid }
        address: { type: string }
        label: { type: string }
        lat: { type: number }
        lng: { type: number }
        activatedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
    DestinationModeState:
      type: object
      required: [active, filtersUsedToday, filtersRemainingToday, config]
      properties:
        active:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/DestinationModeActive'
        filtersUsedToday: { type: integer, minimum: 0 }
        filtersRemainingToday: { type: integer, minimum: 0 }
        disabledUntil:
          type: string
          format: date-time
          nullable: true
          description: When set and in the future, an admin has temporarily disabled destination mode for this driver until this timestamp.
        disabledReason:
          type: string
          nullable: true
          description: Optional admin-provided reason explaining why destination mode is disabled for this driver.
        config: { $ref: '#/components/schemas/DestinationModeConfig' }
    ActivateDestinationModeRequest:
      type: object
      required: [address, lat, lng]
      properties:
        address: { type: string, minLength: 1, maxLength: 300 }
        label: { type: string, maxLength: 80 }
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
        googlePlaceId: { type: string, maxLength: 200, nullable: true }
    DriverSavedPlace:
      type: object
      required: [id, kind, label, address, lat, lng, lastUsedAt]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [home, work, recent] }
        label: { type: string }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        googlePlaceId: { type: string, nullable: true }
        lastUsedAt: { type: integer, format: int64, description: "Epoch ms." }
    DriverSavedPlacesResponse:
      type: object
      required: [home, work, recents]
      properties:
        home:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/DriverSavedPlace'
        work:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/DriverSavedPlace'
        recents:
          type: array
          items: { $ref: '#/components/schemas/DriverSavedPlace' }
    UpsertDriverSavedPlaceRequest:
      type: object
      required: [kind, address, lat, lng]
      properties:
        kind: { type: string, enum: [home, work, recent] }
        label: { type: string, maxLength: 80 }
        address: { type: string, minLength: 1, maxLength: 300 }
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
        googlePlaceId: { type: string, maxLength: 200, nullable: true }
...